Changes "readOnly" properties to required properties.
First, "readOnly" doesn't really mean much in this context, because the
model is not strongly coupled with the database on the backend (i.e.,
changing your object has no impact on the backend until you make a
separate API call to enforce the change).

Second, we want all values to be populated in this API, so "required" is
the most appropriate setting to meet this.

Signed-off-by: Jordan Deyton <jordan.deyton@gmail.com>